+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](http://keepachangelog.com).
 
 ## [Unreleased]
+- (none)
+
+## [2.1.0] - 2016-08-26
 ### Added
 - Added [`no-arguments`] rule
 - Added [`no-rest-parameters`] rule
@@ -46,7 +49,8 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 - Added [`no-this`] rule
 - Added [`no-throw`] rule
 
-[Unreleased]: https://github.com/jfmengels/eslint-plugin-fp/compare/v2.0.0...master
+[Unreleased]: https://github.com/jfmengels/eslint-plugin-fp/compare/v2.1.0...master
+[2.1.0]: https://github.com/jfmengels/eslint-plugin-fp/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/jfmengels/eslint-plugin-fp/compare/v1.3.0...v2.0.0
 [1.3.0]: https://github.com/jfmengels/eslint-plugin-fp/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/jfmengels/eslint-plugin-fp/compare/v1.1.0...v1.2.0